backend/metalit_challenge/requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
